스킬 시스템 구현 중

+++ Assets/Script/SkillScript/SkillData.cs
@@ -8,7 +8,7 @@ public class SkillDatas{
 [Serializable]
 public class SkillData {
     public int skill_type;
-    public int skill_damage;
+    public float skill_damage;
     public float skill_cooltime;
     public string skill_info;
     public float coolDown = 0;
